Forward metadata when publishing messages
One of the key features of the project that Velveteen is being extracted
from is that message metadata is forwarded all the way through the
pipeline. The idea is that the originating message has the context about
where the data belongs (e.g. an id for a database record) and all the
steps throughout the pipeline generally don't care about that context,
until the very end when the data is persisted somewhere. With automatic
forwarding, we can set that context in the beginning and rely on it
being available at the end, without explicitly tracking it all the way.