Add basic error handling
Adds an error handler to the global config. Default value for now prints
the error and discards the message. Whether that makes sense as a
default is to be determined. If so, it'll likely be updated to be more
robust. An upcoming task is to add an error handler with exponential
backoff support.

Also updates the example worker to randomly fail, allowing us to see the
behavior in action.