Move channel and exchange into config
These aren't quite config, but the convenience of doing this is very
high. Almost high enough that it seems to good to be true. It's probably
a good idea to keep an eye on further improvements. E.g. having a
`Config` instance somewhere (`Velveteen.config`?) instead of module
level methods for everything.