Fix flaky spec
The memoization in Config can sometimes cache a mocked object, which can
cause failures based on spec order. This project is not yet far along
enough to worry about performance tweaks, so this commit removes the
memoization of the exchange.

Also updates the Reject handler so it doesn't print to stdout when
running the specs.