First pass at exponential backoff
This takes the approach of dynamically creating retry queues based on
how long the failed messages should be delayed. E.g. a message that has
failed twice already will be sent to a queue with a 9 second TTL (third
attempt is 3^2 = 9). When the message expires from that queue, it goes
back to the original queue.

Adapted from this post: https://m.alphasights.com/exponential-backoff-with-rabbitmq-78386b9bec81