Simplify token production
Updates the token bucket to publish tokens directly to the queue, so it
no longer needs the exchange, which simplifies the rate-limit interface.

Also shifts the terminology from "key" to "queue name" to more clearly
express how it's used and makes a subtle shift from `-` delimiters to
`_` delimiters in the example and specs.