Consistent error handler interface
In c6d479a8 the reject backoff handler was updated to accept a worker
class, instead of an instance, because some errors happen before a
worker is instantiated and resulting in passing in `nil`. That change
left the two error handlers with a slightly different interface, which
is corrected here.