Add worker timeout
If a worker hangs for an extended amount of time, it significantly slow
down or halt a pipeline. One way we can deal with that is to force the
worker to stop after a certain amount of time.

This uses Rub'y builtin Timeout class to cause an exception to be raised
when the limit, defaulted to 10 seconds, is reached.